Use the bower installed as part of the npm install process and check for the existence of the instance/config.py before proceeding with the installation.

+++ dev_install.sh
@@ -2,6 +2,11 @@
 
 echo "This script assumes you have python 2.7, virtualenv, and npm installed. Firefox is required to run the javascript tests"
 
+if [ ! -s instance/config.py ]; then
+   echo "Please create a instance/config.py file before proceeding. See the README.md for what's required"
+   exit
+fi
+
 if [ "$1" == "--clean" ]; then
    echo "Cleaning the project"
    rm -rf node_modules;
@@ -20,7 +25,7 @@ env/bin/pip install -r requirements.txt;
 
 echo "Installing the node and bower dependencies"
 npm install;
-bower install;
+node_modules/bower/bin/bower install;
 
 echo "Successfully installed the pubs_ui app. "
 echo "Run all tests"
@@ -29,9 +34,7 @@ node_modules/karma/bin/karma start tests/js/metrics/karma.conf.js
 env/bin/nosetests -w pubs_ui;
 env/bin/lettuce --with-xunit pubs_ui/lettuce_testing;
 
-if [ ! -s instance/config.py ]; then
-   echo "Please create a instance/config.py file before proceeding. See the README.md for what's required"
-fi
+
 
 echo "When running scripts in this project set up the python interpreter to use env/bin/python."
 echo "The development server can be run as follows: env/bin/python run.py"
